fix: restore paint-to-audio pipeline with RT-safe processing
- Fix coordinate scaling bug (remove /1000.0f division)
- Add visual tracer dots for brush feedback
- Enable debug logging to C:\temp\spectral_debug.log
- Keep Canvas mode active (disable Forge override)
- Boost audio output volume 3x for audibility
- Update notification rules in CLAUDE.md

+++ Source/Core/PaintQueue.h
@@ -4,7 +4,7 @@
 #include <juce_graphics/juce_graphics.h>
 
 struct StrokeEvent { 
-    int x=0,y=0; 
+    float x=0.0f,y=0.0f; // Changed to float to preserve 0-1 coordinate precision
     float pressure=1.f,size=1.f,speed=0.f; 
     juce::Colour colour=juce::Colours::white; 
 };
